qca-wifi-oss: Populate the puncture pattern from the NOL IE
dfs_get_radar_bitmap_from_nolie() API populates the radar bitmap
from the NOL IE bitmap. This radar bitmap is then fed to
ieee80211_dfs_action so that the target channel in the CSA is chosen
as the punctured channel if puncture pattern is valid and DFS puncturing
feature is enabled. The radar frequencies are computed from the NOL IE
bitmap and the NOL start frequency. The bit position of these radar
frequncies in the current channel is found out and these bit positions
are punctured (encoded as 0x1) to create the puncture pattern for radar
channel.

CRs-Fixed: 3524605
Change-Id: I20e7dfb30ec1e217b08dc7410821db97b1ca468e