qca-oss: Don't ignore stop_ap when there are no channels available
In outdoor deployment mode, hostapd sends a stop_ap to driver under 2
conditions:
A) When the new AFC response doesn't have the current operating channel,
it is marked as NO_IR and stop_ap is issued.
B) When there are no channels in the AFC response, all channels are marked
NO_IR and stop_ap is issued.

When stop_ap is received for the first case, it should be ignored if there
are other channels to operate in.

To do this, check if there are any channels available before ignoring
stop_ap sent from hostapd.

Change-Id: Id30c44db8aa5c23b478f8ef3f3eeb093a035b447