qcacld-3.0: Retain BW during CSA due to gDisableDFSChSwitch
If "gDisableDFSChSwitch" ini is set, SAP do not change the
channel even if radar is detected. Instead, the SAP is
expected to do a CSA to the same channel. However, during
this CSA, the bandwidth is downgraded to 20 MHz.

To fix this, retain the current SAP operating channel width
in CSA similar to the channel.

Change-Id: I7fa65765ce5f5d249ec65904749af589679dfe3e
CRs-Fixed: 3538115