qcacld-3.0: Initialize signal field in hdd_sinfo with default value
When aggregated mlo station stats are requested, driver
will get vdev station stats for all the ml links and updates
hdd_sinfo struct, which is then sent to userspace.

But, in the current implementation of the driver, if a vdev
is inactive, we return from the function wlan_hdd_update_mlo_sinfo
without updating the hdd_sinfo struct, and when active vdev tries
to update the signal field it fails because the initial value
for the signal is 0dBm which is very high compared to the link's RSSI.

So to resolve the issue, initialize the signal field with a default
value of -128dBm.

Change-Id: I7b584bf431bda38ed9154f2945c9c81fdfd96902
CRs-Fixed: 3562515