qcacld-3.0: Always initialize num VDEV create in adapter
To identify number of VDEVs to create post mode change,
'num_links_on_create' field in adapter is initialized on
open adapter. However this field is not set to desired
value incase of 11be capability is not present.

This can lead to resetting active links bitmap in adapter
to zero. This bitmap is extensively used for numerous VDEV
related iteration/search operations including VDEV create.

In n-link changes, hdd_adapter_disable_all_links() API is
modified to restore bitmap of active links which might change
as part of link switch.
In stop adapter, this API is called without check for eht_cap
and resulting in zero active link bitmap.

Irrespective of 11be capability fill the 'num_links_on_create'
to approripate value so that further usage of this field will
not have undesirable results.

Change-Id: I6dddc8bb57b012dc480e05653db50f6cafb26ccf
CRs-Fixed: 3575089