qca-wifi: Modify the conditions for a txable channel
The API reg_is_6ghz_freq_txable() when called with 5 GHz frequency or with
VLP power mode in outdoor mode will provide incorrect results. Modify the
API as follows,
i) For non-6 GHz frequencies, the API should check if the frequency is
enabled or not.
ii) For Indoor deployment, the API should check if the frequency is enabled
or not.
iii) For outdoor deplyment, the API should check the power mode. Based on
which if the power mode is,
a) VLP, then check if the frequency is enabled or not.
b) SP, then check if AFC is done or not.
iv) For non-6 GHz version of this API, the definition should always return
true.

The AFC done check should be done on super channel entry instead of current
channel list because the current channel list may not be updated
immediately but the super channel list is updated immediately after an AFC
response.

This API reg_is_freq_txable() is moved from cmndev to componentdev by the
change I1031a20d40ebcf30326c14901d413669206a8433.

Change-Id: I36eff5a49b0a274135d0001bba3b5a4f1cc52b3e
CRs-Fixed: 3584154