qcacld-3.0: Remove HDD wait for ops in link switch notify
VDEV is moved to INIT state as part of link switch disconnect,
before set MAC address response is received, north bound
disconnect is received which also increment pending OSIF
ops.

MLO manager is notified about this userspace disconnect to
terminate link switch and in this process, OSIF/HDD is notified
to change back deflink by waiting for OSIF ops to complete on
that adapter. Deadlock situation araises as the same thread
that initiated ops is waiting for completion.

Don't wait for ops in OSIF as part of link switch notification.

Change-Id: I53ce3da7e9cfbc1dea9422766f81c33254b357a0
CRs-Fixed: 3591342