[component-dev] Mesh agent AWGN event notification support
Currently when AWGN is detected by an agent (repeater)
but not by controller, the event is not communicated
to the controller. However, agent acts on AWGN and
moves the BH link to a different channel.
As the controller is not aware of this, this would
result in connection loss (in MLO case, one of the
links becomes unusable).This FR shall address this by
1. Sending AWGN detected event to EasyMesh agent.
2. Handling AWGN found command from EasyMesh agent.
On AWGN found event, trigger AWGN channel change event.

CRs-Fixed: 3606629
Change-Id: I6c288ec15ea657297cbb2ce658e2107a4e268e7d
(cherry picked from commit c39fff6ce6a1c3ffee715862420c9b185532141b)