qcacld-3.0: Select IPA TX pipe based on connection details
This change is one part of patch series to address IPA TX pipe
selection on sdxpinn platform, which is attached with 11BE/WiFi-7
capable chipset. One of the typical user scenarios are the SBS/HBS
configuration supported by the WiFi chipset.

From WLAN point of view, a pipe stands for a pair of TX and TX
completion ring resources. Starting from DBS capable chipset on
SDX platforms, two pipes are allocated for IPA usage, which are the
primary pipe and the alternate pipe. By design, the primary pipe
is used for 5 GHz/6 GHz connection while the alternate pipe is
for 2 GHz connection. Due to different throughput requirements
between two bands and memory saving considerations, two pipes are
initialized with different ring sizes as well as number of buffers
available, where the primary pipe has larger ring sizes as well as
number of TX buffers.

Starting from 11be/WiFi-7 chipset, SBS/HBS features are supported.
Take SAP-SAP concurrencies for instance, SAP1 can be configured to
run with channel 36 on 5 GHz while SAP2 can be configured to run with
channel 149 on 5 GHz or even in the 6 GHz band. With legacy DBS solution,
both SAP1 and SAP2 will use the primary pipe, which in turns TX
starvations are observed on the SAP which has higher phyrate.

To resolve such TX starvation issue, this change is to properly
select TX pipes for interfaces that are possible to be run
on 2 GHz/5 GHz/6 GHz bands but with below considerations being
considered as of the implementation time.
a. 3 home channels are not possible.
b. at same time, only two bands are possible.
c. two 11be SAP concurrencies are not possible.
d. 11be 320 MHz is only possible for single STA or single SAP.

This means that at one time with a maximum of 4 interfaces,
there must be two interfaces with same frequency.

Hence this change properly selects the TX pipe between the
primary pipe and the alternate pipe for all concurrencies that are
supported.

Change-Id: I60453ce0dae5a67f1af9304beeebaab9189c6896
CRs-Fixed: 3599453