qcacmn: Make array access explicit through pointer arithmetic
In API 'reg_get_partial_afc_req_info' the structure
wlan_afc_host_partial_request is accessed for filling the AFC
request. For filling this request, the successive elements are
indexed as arrays. Some static code analysis tools treat this
as an error and out of bounds violation.

This access is intentional and is not an error. To make the access
explicit, use pointer arithmetic for accessing the successive
elements.

Change-Id: I6de3e05d0fa67ecbe9f4e624153f8bae1a616301