Revert "qcacmn: Protect Reconfig ML draft 3.0 definitions under compile flag"
This reverts most parts of Change-Id
I459867e87c27310cb9b1a200b802ff2dc11e93f4.

Reason for revert: IEEE802.11be draft 2.0 support has been deprecated.
Revert the commit in order to remove compile time support for 11be draft
2.0 based Reconfiguration Multi-Link element protocol definitions. This
is a planned reversal since the compile time support was intended to be
a temporary facility for assisting in the transition to 11be draft 3.0.

A minor modification in the documentation for max length of STA Info
field in Per-STA Profile subelement in Reconfiguration ML IE is
retained so that the usage of the macro is represented in the
documentation.

CRs-Fixed: 3618113
Change-Id: I2e98c2ae89baadeae2d99092a488fcb58525b247